+++ HISTORY.md
@@ -42,6 +42,7 @@
 * adjusted Brazilian Portuguese translation for variadic comparison ops
 * lists: tweaked distribution()
 * German translation update for list "distribution"
+* threads: tweaked "unicode of"
 
 ### 2023-02-26
 * threads: fixed #3183 - determining the length of huge strings fails